use fakejson to aviod package.json update

+++ utils/fake-pkg-json.js
@@ -1,18 +1,22 @@
 const path = require('path')
 const fs = require('fs')
 const isObject = require('lodash').isObject
+const merge = require('lodash').merge
 
 const writeFile = fs.writeFileSync
+const rename = fs.renameSync
 const isExist = fs.existsSync
-const deleteFile = fs.unlink
+const deleteFile = fs.unlinkSync
+
+const timestamp = +new Date
+
+const log = console.log
 
 function formatContent (content) {
   return JSON.stringify(content, null, 2)
 }
 
 function FakeJson (fakeContent, fileName) {
-  const timestamp = +new Date
-
   if (!fakeContent) {
     throw new Error('fake json needs content')
   }
@@ -21,30 +25,42 @@ function FakeJson (fakeContent, fileName) {
     throw new TypeError('fake json needs object as content')
   }
 
+  this.timestamp = timestamp
+
   fileName = fileName && fileName.replace(/.json$/, '') || 'package'
 
   this.fakeContent = fakeContent
   this.jsonPath = path.join(process.cwd(), `${fileName}.json`)
-  this.tmpRealPath = `${this.jsonPath}__${timestamp}`
+  this.tmpRealPath = `${this.jsonPath.replace(/.json$/, '')}__${timestamp}__.json`
+
+  this._keepRealJson()
+  this._create()
 }
 
-FakeJson.prototype.keepRealJson = function () {
+FakeJson.prototype._keepRealJson = function () {
   const realJson = require(this.jsonPath)
 
-  writeFile(this.tmpRealPath, formatContent(realJson))
+  this.realScripts = realJson.scripts
+
+  rename(this.jsonPath, this.tmpRealPath)
+  log()
+  log(`rename ${this.jsonPath} to ${this.tmpRealPath}`)
+  log()
 }
 
-FakeJson.prototype.create = function () {
+FakeJson.prototype._create = function () {
   if (isExist(this.tmpRealPath)) {
-    writeFile(this.jsonPath, formatContent(this.fakeContent))
+    writeFile(this.jsonPath, formatContent(merge({
+      scripts: this.realScripts
+    }, this.fakeContent)))
   } else {
     throw new Error('no real json found. please call keepRealJson first!')
   }
 }
 
-FakeJson.prototype.removeFake = function () {
-  if (isExists(this.tmpRealPath)) {
-    deleteFile(this.jsonPath)
+FakeJson.prototype._removeFake = function () {
+  if (isExist(this.tmpRealPath)) {
+    isExist(this.jsonPath) &&  deleteFile(this.jsonPath)
   } else {
     throw new Error('no real json found. please call keepRealJson first!')
   }
@@ -53,5 +69,11 @@ FakeJson.prototype.removeFake = function () {
 FakeJson.prototype.resetRealJson = function () {
   const realJson = require(this.tmpRealPath)
 
-  writeFile(this.jsonPath, formatContent(realJson))
+  this._removeFake()
+  rename(this.tmpRealPath, this.jsonPath)
+  log()
+  log(`rename ${this.tmpRealPath} to ${this.jsonPath}`)
+  log()
 }
+
+module.exports = FakeJson
